requirements.txt
beautifulsoup4==4.6.0
certifi==2017.11.5
chardet==3.0.4
Django==2.0
django-cors-headers==2.1.0
django-debug-toolbar==1.9.1
django-environ==0.4.4
django-extensions==1.9.8
django-model-utils==3.0.0
django-rest-framework==0.1.0
djangorestframework==3.7.3
idna==2.6
olefile==0.44
Pillow==4.3.0
pymongo==3.6.0
pytesseract==0.1.7
pytz==2017.3
requests==2.18.4
six==1.11.0
sqlparse==0.2.4
typing==3.6.2
urllib3==1.22
Werkzeug==0.13
xid==1.0.4
